+++ src/main/java/com/nsa/cubric/application/services/AdminServicesStatic.java
@@ -2,6 +2,6 @@ package com.nsa.cubric.application.services;
 
 public interface AdminServicesStatic {
     public boolean removeUser(Long userId);
-    public Long removeUserResponses(Long userId);
+    public Integer removeUserResponses(Long userId);
 
 }
